fix(event-streams): address review findings
